feat(club): Enhance ClubDto and add members count
This commit enhances the `ClubDto` object and adds a `members` field representing the number of members in a club. It also includes a debug print statement in `ClubServiceImpl` for inspecting the generated `ClubDto` list.  Additionally, the events field in `ClubDto` is now a list of `EventDto` objects.

The changes include:

- Adding a `members` field to the `ClubDto` object.
- Modifying the `ClubMapper` to populate the `members` and `events` fields in `ClubDto`.
- Adding a `System.out.println()` statement in `ClubServiceImpl.findAllClubs()` to log the list of `ClubDto` objects.
- Modifying the `Club` model to include a `members` field.
- Changing the `events` field type in `ClubDto` from `List<Event>` to `List<EventDto>`.
- Updating the `ClubMapper` to map `Event` objects to `EventDto` objects for the `events` field.

This enhancement provides more information about clubs when retrieved, specifically the number of members and event details, improving the overall data representation.

The debug print statement will be removed in a future commit after thorough testing.