feat(user): Add ManyToMany relationship between User and Club
This commit introduces a ManyToMany relationship between the `User` and `Club` entities.  Users can now belong to multiple clubs, and clubs can have multiple members. This allows for a more flexible and realistic representation of user participation in different clubs.

The `clubs` field has been added to the `UserEntity` with the `@ManyToMany` annotation and the `mappedBy` attribute set to "users". This designates the `users` field in the `Club` entity as the owner of the relationship.

This change lays the groundwork for future features involving user-club interactions, such as club membership management and personalized club content.